1573 part 2: same fix applied to shipyard prices.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@1574 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Universe.m
@@ -7263,8 +7263,8 @@ double estimatedTimeForJourney(double distance, int hops)
 				price = base_price + cunningFee(price - base_price);
 			}
 				
-			[description appendFormat:DESC(@"shipyard-selling-price-d-credits"), price];
-			[short_description appendFormat:DESC(@"shipyard-price-d-credits"), price];
+			[description appendFormat:DESC(@"shipyard-selling-price-d-credits"), (int) price];
+			[short_description appendFormat:DESC(@"shipyard-price-d-credits"), (int) price];
 
 			NSString* ship_id = [NSString stringWithFormat:@"%06x-%06x", super_rand1, super_rand2];
